Moves project structure to the cocoapods default
