+++ Example/Tests/Tests.swift
@@ -1,0 +1,36 @@
+//
+//  Tests.swift
+//  Tests
+//
+//  Created by Christopher DeOrio on 8/30/17.
+//  Copyright © 2017 Christopher DeOrio. All rights reserved.
+
+import UIKit
+import XCTest
+import OTOnexus
+
+class Tests: XCTestCase {
+    
+    override func setUp() {
+        super.setUp()
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+    
+    override func tearDown() {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        super.tearDown()
+    }
+    
+    func testExample() {
+        // This is an example of a functional test case.
+        XCTAssert(true, "Pass")
+    }
+    
+    func testPerformanceExample() {
+        // This is an example of a performance test case.
+        self.measure() {
+            // Put the code you want to measure the time of here.
+        }
+    }
+    
+}
